refactor(maxwell-mobile): extract sub-components to satisfy 500-line cap
`frontend/src/pages/Maxwell/Mobile.tsx` was 901 lines, far over the
500-line soft cap enforced by `ci-standard.yml`. Decomposed into:

- mobileTypes.ts        — shared types, constants, helpers
- TaskCard.tsx          — single task card
- ChatBubble.tsx        — single chat message bubble
- MaxwellStatusHeader.tsx — status pill + refresh/controls buttons
- MaxwellTasks.tsx      — task list with loading / empty states
- MaxwellChat.tsx       — chat composer, history, voice input, chips
- MaxwellControlSheet.tsx — Start / Stop / Restart bottom sheet

Main `Mobile.tsx` (now 399 lines) keeps state, data fetching, chat
streaming, and daemon control logic, and composes the sub-components.
Public export `MaxwellMobile` and all data-testid hooks are preserved.

No behavior change.
